requirements.txt
openai
qdrant-client
pypdf
python-dotenv

